+++ tests/crypto-browser.js
@@ -1,3 +1,2 @@
 // WebCrypto
-/* eslint-env browser */
-export default (globalThis.crypto || globalThis.msCrypto);
+export default globalThis.crypto;