+++ tests/util.js
@@ -1,3 +1,0 @@
-export {
-  ReadableStream, TransformStream
-} from 'web-streams-polyfill/dist/ponyfill.mjs';